+++ src/pyws/adapters/_twisted.py
@@ -4,12 +4,12 @@ def serve(request, server):
     """
     Twisted Web adapter. It has two arguments:
 
-    #. a Twisted Web request object,
-    #. a pyws server object.
+    #. ``request`` is a Twisted Web request object,
+    #. ``server`` is a pyws server object.
 
     First one is the context of an application, function ``serve`` transforms
     it into a pyws request object. Then it feeds the request to the server,
-    gets a response, sets header ``Content-Type`` and returns response text.
+    gets the response, sets header ``Content-Type`` and returns response text.
     """
 
     request_ = Request('/'.join(request.postpath),